[CSL-1081] Move block types into Pos.Block.Core
