+++ src/Pos/Block/Core.hs
@@ -1,0 +1,3 @@
+-- | Core of block processing.
+
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}
